// src/sdk/model/btc-single-address-root.ts
/**
 * UnBlock RESTful Web API
 * An blockchain analysis service
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BtcSingleAddressRoot { 
    address?: string;
    endingTime?: Date;
    maxLevel?: number;
    startingTime?: Date;
    timeoutSecs?: number;
}